pyproject.toml
[tool.poetry]
name = "qlik-cloud-console2"
version = "0.1.0"
description = ""
authors = ["rodrigo ubuntu <digaosss70@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.32.2"
python-dotenv = "^1.0.1"
streamlit-option-menu = "^0.3.12"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
